performed automatic code improvement (#21)
* performed automatic code improvement

* fixes after unit tests failure

* for loops and boolean parsing

* use diamond notation

* use try with resources

* use utility comparision

* use switch

* use lambda

* make inner class static

* numeric fixes

* don't use pre sized array

* some formatting

* fine tune lint checks

* use apply instead of commit for the preferences, some XML corrections

+++ app/src/main/java/ru/orangesoftware/financisto/graph/GraphWidget.java
@@ -60,8 +60,8 @@ public class GraphWidget extends View {
 			canvas.drawRect(x, y, x+lineWidth, y+style.lineHeight, style.linePaint);
 			style.amountPaint.setColor(amount == 0 ? zeroColor : (amount > 0 ? positiveColor : negativeColor));
 			canvas.drawText(a.getAmountText(), 
-					x+lineWidth+style.textDy+a.amountTextWidth/2, 
-					y+style.lineHeight/2+style.amountHeight/2, 
+					x+lineWidth+style.textDy+ (float) a.amountTextWidth /2,
+					y+ (float) style.lineHeight /2+ (float) style.amountHeight /2,
 					style.amountPaint);
 			y += style.lineHeight+style.dy;
 		}
