+++ app/src/main/java/ru/orangesoftware/financisto/widget/AmountPicker.java
@@ -52,7 +52,7 @@ public class AmountPicker extends LinearLayout implements NumberPicker.OnChanged
         	TextView dotText = new TextView(context);
         	dotText.setText(".");
         	dotText.setGravity(Gravity.BOTTOM);
-        	dotText.setTextSize(pickerWidth/2);
+        	dotText.setTextSize((float) pickerWidth /2);
         	addView(dotText, lpWrapFill);
         	addPickers(context, lpFixedWrap, fractions);
         }
